+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes
 - Find and use magik-lint.properties based on file path, instead of just once, in magik-language-server.
 - Add --enabled=... and --disabled=... to magik-lint; use --disabled=all with --enable=... to enable specific checks.
 - Add type hierarchy provider to magik-language-server.
+- Add inlay hint provider to magik-language-server.
 - Various small fixes.